+++ src/localization/slam_marrano/src/icp_slam.cpp
@@ -20,6 +20,7 @@ using namespace std;
 
 ICP_handle::ICP_handle(){
 	previous_map = pcl::PointCloud<PointXYZColorScore>::Ptr(new pcl::PointCloud<PointXYZColorScore>);
+	allp_clustered = pcl::PointCloud<PointXYZColorScore>::Ptr(new pcl::PointCloud<PointXYZColorScore>);
 
 	position = Eigen::Matrix4f::Identity(4, 4);
 	prev_transformation = Eigen::Matrix4f::Identity(4, 4);
@@ -33,12 +34,12 @@ ICP_handle::ICP_handle(){
 }
 
 void ICP_handle::map_callback(sensor_msgs::PointCloud2 map_msg) {
-	callback_iteration++;
 
 	pcl::PointCloud<PointXYZColorScore>::Ptr new_map(new pcl::PointCloud<PointXYZColorScore>);
 	pcl::fromROSMsg(map_msg, *new_map);
 
 	if(!has_map){
+		*allp_clustered = *new_map;
 		*previous_map = *new_map;
 		has_map = true;
 
@@ -52,7 +53,10 @@ void ICP_handle::map_callback(sensor_msgs::PointCloud2 map_msg) {
 
 	pcl::IterativeClosestPoint<PointXYZColorScore, PointXYZColorScore> icp;
 	icp.setInputSource(map_in_position);
-    icp.setInputTarget(previous_map);
+        if(callback_iteration < 10)
+            icp.setInputTarget(previous_map);
+        else
+            icp.setInputTarget(allp_clustered);
 	icp.setMaximumIterations(5);
 	icp.setEuclideanFitnessEpsilon (0.005);
 	icp.setTransformationEpsilon(1e-5);
@@ -66,128 +70,122 @@ void ICP_handle::map_callback(sensor_msgs::PointCloud2 map_msg) {
 
 	float ang = (float)-atan2(transformation.coeff(0, 1), transformation.coeff(0,0));
 	float dist = pow((transformation.coeff(0,3) - prev_transformation.coeff(0,3)),2) + pow((transformation.coeff(1,3) - prev_transformation.coeff(1,3)),2) + pow((transformation.coeff(2,3) - prev_transformation.coeff(2,3)),2);
-	// std::cout << dist << std::endl;
+
+	dist_media = ((callback_iteration-1)*(dist_media)+dist)/callback_iteration;
+
+	std::cout << dist << "    " << dist_media << std::endl;
+
+	// if(dist>0.5){
+	// 	sensor_msgs::PointCloud2 new_map_msg;
+
+	// 	pcl::toROSMsg(*allp_clustered, new_map_msg);
+	// 	new_map_msg.header.frame_id = "map";
+	// 	map_publisher.publish(new_map_msg);
+
+	// 	*previous_map += *previous_map;
+
+	// 	return;
+	// }
 	prev_transformation = transformation;
 
 
 	position = transformation * position;
 	send_position();
 
+	*previous_map += registered_map;
 
-	//*previous_map += registered_map;
-	for(PointXYZColorScore new_point : registered_map.points){
-		pcl::PointCloud<PointXYZColorScore> distances_cloud;
-		float min_dist=INFINITY;
-		int index;
-		for(int i=0 ; i<previous_map->points.size(); i++){
-			float dist = sqrt(pow(previous_map->points[i].x-new_point.x,2)+pow(previous_map->points[i].y-new_point.y,2));
-			if(dist<min_dist){
-				min_dist = dist;
-				index = i;
+
+	pcl::search::KdTree<PointXYZColorScore>::Ptr tree (new pcl::search::KdTree<PointXYZColorScore>);
+	tree->setInputCloud(previous_map);
+
+	std::vector<pcl::PointIndices> cluster_indices;
+	pcl::EuclideanClusterExtraction<PointXYZColorScore> ec;
+	ec.setClusterTolerance (1.5);
+	ec.setMinClusterSize (3);
+	ec.setMaxClusterSize (100000);
+	ec.setSearchMethod (tree);
+	ec.setInputCloud (previous_map);
+	ec.extract (cluster_indices);
+
+	pcl::PointCloud<PointXYZColorScore>::Ptr clustered_points (new pcl::PointCloud<PointXYZColorScore>);
+	for(const auto& cluster : cluster_indices) {
+		PointXYZColorScore centro;
+		centro.x = 0;
+		centro.y = 0;
+		centro.z = 0;
+		centro.color = 0;
+		centro.score = 1;
+		for (const auto& idx : cluster.indices) {
+			centro.x += (*previous_map)[idx].x;
+			centro.y += (*previous_map)[idx].y;
+			centro.z = 0;
+
+		}
+		centro.x /= cluster.indices.size();
+		centro.y /= cluster.indices.size();
+		for (const auto& idx : cluster.indices) {
+			(*previous_map)[idx].x = centro.x;
+			(*previous_map)[idx].y = centro.y;
+		}
+
+		clustered_points->push_back(centro);
+
+	}
+	*allp_clustered = *clustered_points;
+
+
+
+	if(callback_iteration % 10 == 9) {
+		std::vector<pcl::PointIndices> cluster_indices;
+		pcl::EuclideanClusterExtraction<PointXYZColorScore> ec;
+		ec.setClusterTolerance (1.5);
+		ec.setMinClusterSize (6);
+		ec.setMaxClusterSize (200);
+		ec.setSearchMethod (tree);
+		ec.setInputCloud (previous_map);
+		ec.extract (cluster_indices);
+
+
+
+		pcl::PointCloud<PointXYZColorScore>::Ptr clustered_points (new pcl::PointCloud<PointXYZColorScore>);
+		pcl::PointCloud<PointXYZColorScore>::Ptr mapa_global (new pcl::PointCloud<PointXYZColorScore>);
+		for(const auto& cluster : cluster_indices) {
+			int j = 0;
+			PointXYZColorScore cono;
+			cono.x = (*previous_map)[cluster.indices[0]].x;
+			cono.y = (*previous_map)[cluster.indices[0]].y;
+			cono.z = 0;
+			cono.color = 0;
+			cono.score = 1;
+			for (const auto& idx : cluster.indices) {
+                                // This is a botch that works but should be changed. PCL ICP doesn't sopport weighted points,
+                                // so the more reliable we consider a point to be, the more that are placed on the exact same
+                                // coordinates (limited in this case to 50).
+				if(j < 10)
+					clustered_points->push_back((*previous_map)[idx]);
+				else
+					break;
+				j++;
 			}
-
-		}
-		std::cout << min_dist << std::endl;
-		if(min_dist<0.5){
-			previous_map->points[index].x = 0.9*previous_map->points[index].x + 0.1*new_point.x;
-			previous_map->points[index].y = 0.9*previous_map->points[index].y + 0.1*new_point.y;
-		} else {
-			previous_map->push_back(new_point);
+			mapa_global->push_back(cono);
 		}
 
+		// sensor_msgs::PointCloud2 map_msg;
+
+		// pcl::toROSMsg(*mapa_global, map_msg);
+		// map_msg.header.frame_id = "map";
+		// map_publisher.publish(map_msg);
+
+		*previous_map = *clustered_points;
 	}
 
 	sensor_msgs::PointCloud2 new_map_msg;
 
-	pcl::toROSMsg(*previous_map, new_map_msg);
+	pcl::toROSMsg(*allp_clustered, new_map_msg);
 	new_map_msg.header.frame_id = "map";
 	map_publisher.publish(new_map_msg);
 
-	// pcl::search::KdTree<PointXYZColorScore>::Ptr tree (new pcl::search::KdTree<PointXYZColorScore>);
-	// tree->setInputCloud(previous_map);
-
-	// std::vector<pcl::PointIndices> cluster_indices;
-	// pcl::EuclideanClusterExtraction<PointXYZColorScore> ec;
-	// ec.setClusterTolerance (1.5);
-	// ec.setMinClusterSize (3);
-	// ec.setMaxClusterSize (100000);
-	// ec.setSearchMethod (tree);
-	// ec.setInputCloud (previous_map);
-	// ec.extract (cluster_indices);
-
-	// pcl::PointCloud<PointXYZColorScore>::Ptr clustered_points (new pcl::PointCloud<PointXYZColorScore>);
-	// for(const auto& cluster : cluster_indices) {
-	// 	PointXYZColorScore centro;
-	// 	centro.x = 0;
-	// 	centro.y = 0;
-	// 	centro.z = 0;
-	// 	centro.color = 0;
-	// 	centro.score = 1;
-	// 	for (const auto& idx : cluster.indices) {
-	// 		centro.x += (*previous_map)[idx].x;
-	// 		centro.y += (*previous_map)[idx].y;
-	// 		centro.z = 0;
-
-	// 	}
-	// 	centro.x /= cluster.indices.size();
-	// 	centro.y /= cluster.indices.size();
-	// 	for (const auto& idx : cluster.indices) {
-	// 		(*previous_map)[idx].x = centro.x;
-	// 		(*previous_map)[idx].y = centro.y;
-	// 	}
-
-	// 	clustered_points->push_back(centro);
-
-	// }
-	// *allp_clustered = *clustered_points;
-
-
-
-	// if(callback_iteration % 10 == 9) {
-	// 	std::vector<pcl::PointIndices> cluster_indices;
-	// 	pcl::EuclideanClusterExtraction<PointXYZColorScore> ec;
-	// 	ec.setClusterTolerance (1.5);
-	// 	ec.setMinClusterSize (6);
-	// 	ec.setMaxClusterSize (200);
-	// 	ec.setSearchMethod (tree);
-	// 	ec.setInputCloud (previous_map);
-	// 	ec.extract (cluster_indices);
-
-
-
-	// 	pcl::PointCloud<PointXYZColorScore>::Ptr clustered_points (new pcl::PointCloud<PointXYZColorScore>);
-	// 	pcl::PointCloud<PointXYZColorScore>::Ptr mapa_global (new pcl::PointCloud<PointXYZColorScore>);
-	// 	for(const auto& cluster : cluster_indices) {
-	// 		int j = 0;
-	// 		PointXYZColorScore cono;
-	// 		cono.x = (*previous_map)[cluster.indices[0]].x;
-	// 		cono.y = (*previous_map)[cluster.indices[0]].y;
-	// 		cono.z = 0;
-	// 		cono.color = 0;
-	// 		cono.score = 1;
-	// 		for (const auto& idx : cluster.indices) {
-    //                             // This is a botch that works but should be changed. PCL ICP doesn't sopport weighted points,
-    //                             // so the more reliable we consider a point to be, the more that are placed on the exact same
-    //                             // coordinates (limited in this case to 50).
-	// 			if(j < 10)
-	// 				clustered_points->push_back((*previous_map)[idx]);
-	// 			else
-	// 				break;
-	// 			j++;
-	// 		}
-	// 		mapa_global->push_back(cono);
-	// 	}
-
-	// 	// sensor_msgs::PointCloud2 map_msg;
-
-	// 	// pcl::toROSMsg(*mapa_global, map_msg);
-	// 	// map_msg.header.frame_id = "map";
-	// 	// map_publisher.publish(map_msg);
-
-	// 	*previous_map = *clustered_points;
-	// }
-
-
+	callback_iteration++;
 
 }
